select2

+++ app/Http/Controllers/CategoriesController.php
@@ -99,6 +99,10 @@ class CategoriesController extends Controller
     public function destroy(Category $category)
     {
         //
+        if($category->posts->count() > 0 ){
+            session()->flash('error','Category can not be Deleted ');
+            return redirect()->back();
+        }
         $category->delete();
         session()->flash('success','Category Deleted successfully');
         return redirect(route('categories.index'));
